(backend): Many updates - Created fields for instagram, facebook and twitter on posts - Manage file upload for post - post_when field is required now - Changed posts services to work for the changes above

+++ backend/app/Services/Post/DeleteService.js
@@ -1,10 +1,15 @@
 'use strict'
 
 const Post = use('App/Models/Post')
+const GeneralException = use('App/Exceptions/GeneralException')
 
 class UpdateService {
   async handle(id) {
     const post = await Post.findOrFail(id)
+
+    if (post.posted_when) {
+      throw new GeneralException(400, "Can't delete a posted post")
+    }
 
     await post.delete()
   }
